Fixed blacklist terms score parsing bug. Re-enabled anti-spammer abuse mechanism. Tightened command checks. Updated filters.

+++ Phamhilator/BlackFilters.Offensive.cs
@@ -53,7 +53,7 @@ namespace Phamhilator
 						continue;
 					}
 
-					Terms.Add(term, int.Parse(termScore));
+					Terms.Add(term, float.Parse(termScore));
 				}
 			}
