+++ Phamhilator/BlackFilters.Spam.cs
@@ -47,7 +47,7 @@ namespace Phamhilator
 
 					if (Terms.ContainsTerm(term)) { continue; }
 
-					Terms.Add(term, int.Parse(termScore));
+					Terms.Add(term, float.Parse(termScore));
 				}
 			}
